+++ kubectl/cmd.py
@@ -72,8 +72,11 @@ def logsPod(podName,namespace,options):
     output = executeCmd(cmd)
     return output
 
-def getNodes():
+def getNodes(noderole=None):
+    #kubectl get nodes -l node-role.kubernetes.io/worker=true
     cmd="kubectl get nodes "
+    if noderole != None:
+       cmd = "%s -l node-role.kubernetes.io/%s=true" % (cmd,noderole)
     output = executeCmd(cmd+" --no-headers")
     return output
 
@@ -82,14 +85,27 @@ def describeNode(nodeName):
     output = executeCmd(cmd)
     return output
 
-def getPods(namespace):
+def getPods(namespace,nodeNameList=[]):
     cmd="kubectl get pods "
+
     if namespace == "all-namespaces":
         cmd=cmd+"--"+namespace
     else:
         cmd=cmd+"-n "+namespace
     cmd=cmd+" -o wide "
-    output = executeCmd(cmd+" --no-headers")
+    cmd=cmd+" --no-headers"
+    output = ""
+    if nodeNameList != None and len(nodeNameList)>0:
+        #get pods for specified nodes
+        for nodeName in nodeNameList:
+            #kubectl get pods --all-namespaces  --no-headers --field-selector spec.nodeName=10.31.10.126    
+            cmd2="%s --field-selector spec.nodeName=%s" % (cmd,nodeName)
+            output2 = executeCmd(cmd2)
+            if output2.lower().find("no resources found") == -1:
+                output = output + output2
+    else:
+        output = executeCmd(cmd)
+
     return output
 
 def getNamespaces():