+++ kubectl/nodes.py
@@ -11,6 +11,15 @@ def describe(cmdOptions, selectedNode):
         return "Describing all nodes not (yet) supported."
     return describeNode(node)
     
+def getWorkerNodeNames():
+    workerNodes = getNodes(noderole="worker")
+    nodeNames = []
+    for node in workerNodes.split("\n"):
+        if node != "":
+            nodeFields=node.split()
+            nodeName=nodeFields[0]
+            nodeNames.append(nodeName)
+    return nodeNames
 
 def list():
     '''Return list of tuples of nodes: [(value,label),(value,label),...]'''
@@ -27,9 +36,8 @@ def list():
             value="%s %s %s %s" % (nodeFields[0],readyString,nodeFields[2],nodeFields[4])
             nodes.append((nodeFields[0],value))
 
-    nodes.insert(0,("all","all, ready %d/%d" % (nodesReady,len(nodes)) ))
-    #return values for RadioList
+    nodes.insert(0,("workers","all worker nodes"))
+    nodes.insert(0,("all","all, ready %d/%d" % (nodesReady,len(nodes)-1) ))
+
     return nodes
     
-    #return multiline string
-    #return allNodes.strip()